pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "phoenix-ai"
version = "0.1.0"
description = "AI-powered real estate agent for outbound calling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit>=1.0",
    "livekit-agents[openai,deepgram,cartesia,silero,google,groq]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-turn-detector~=1.2",
    "python-dotenv~=1.0",
    "torch>=2.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
